Do not terminate if duplicate tx is received. Just ignore it.
Duplicate transactions happen usually when somebody posts
extraordinary large transaction to the bitcoin network and it is not
confirmated at all or takes very long time. In that time our trusted
peer flushes the transaction from its memory pool. Our pool, on the
contrary, is infinitive and we can see them twice.

Another way of fixing this is to ignore duplicates when responding to
inv requests. It would be relatively easy to fix it there, too.