Add compacted message headers
This patch makes it possible to store arbitrary amounts of data in
message heading which is useful for storing transaction ages, etc.

This has bit ugly side-effect of making incoming_node.c more
complex. Because we want to handle inv requests efficiently, we are
using a bit dangerous pointer casting techniques to achieve maximimal
memory utilization.