Make poll() FIFO compatible and make core dump if poll results are weird
Reading from FIFOs when the feeder process were dead made the
deserializer go wild and consume 100% CPU. To fix this, a new command
line option (-w) is added. To prevent these kind of programming errors
in the future, I added a worst case handler which terminates the
program (and allows core dump to happen).

Thanks to Steven Spark and mark4o in Stack Overflow for the trick:

http://stackoverflow.com/questions/22021253/poll-on-named-pipe-returns-with-pollhup-constantly-and-immediately