Fix an issue when a block gets missing in the block explorer #18
For some reasons I'm not aware of, the FIMK block explorer
sporadically destroys some URLs it has already given links to. While
this is bad, let it not break the whole streaming and we just jump
over the missing ones.

My guess is that the FIMK blockchain splits so often because of 30
second block time and the block explorer tries to be clever and
destroys them from the whole history (which is definitely not the way
to go...)